pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "overlay"
version = "0.1.0"
description = "Agentic image generation + smart text overlay"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.92.0",
    "openai>=1.50",
    "pydantic>=2.0",
    "Pillow>=10.0",
    "diffusers>=0.30",
    "torch>=2.2",
    "transformers>=4.40",
    "accelerate>=0.30",
    "safetensors>=0.4",
]

[project.scripts]
overlay = "overlay.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["overlay*"]

[tool.setuptools.package-data]
overlay = ["fonts/*.ttf", "fonts/*.otf"]
